requirements.txt
dash==2.14.2
dash-cytoscape==0.3.0
plotly==5.17.0
pandas==2.1.4
sqlalchemy==2.0.23
google-generativeai==0.3.2
python-dotenv==1.0.0
flask==3.0.0
pyperclip==1.8.2
